+++ lib/kettle/dev.rb
@@ -49,6 +49,7 @@ module Kettle
     autoload :ReleaseNotifier, "kettle/dev/release_notifier"
     autoload :ReleaseCLI, "kettle/dev/release_cli"
     autoload :ReleaseSecrets, "kettle/dev/release_secrets"
+    autoload :ReleaseSecretsDoctor, "kettle/dev/release_secrets_doctor"
     autoload :ResetCLI, "kettle/dev/reset_cli"
     autoload :PreReleaseCLI, "kettle/dev/pre_release_cli"
     autoload :Version, "kettle/dev/version"